+++ sklearn/preprocessing.py
@@ -1,9 +1,12 @@
 # Authors: Alexandre Gramfort <alexandre.gramfort@inria.fr>
 #          Mathieu Blondel <mathieu@mblondel.org>
 #          Olivier Grisel <olivier.grisel@ensta.org>
+#          Andreas Mueller <amueller@ais.uni-bonn.de>
 # License: BSD
+
 from collections import Sequence
 import warnings
+import numbers
 
 import numpy as np
 import scipy.sparse as sp
@@ -600,6 +603,121 @@ def _is_multilabel(y):
     return not isinstance(y[0], np.ndarray) and isinstance(y[0], Sequence) \
        and not isinstance(y[0], basestring) \
         or _is_label_indicator_matrix(y)
+
+
+class OneHotEncoder(BaseEstimator, TransformerMixin):
+    """Encode categorial integer features using a one-hot scheme.
+
+    The input is assumed to be integer and name categorial
+    features numbered from ``0`` to ``n_values - 1``.
+
+    Parameters
+    ----------
+    n_values : 'auto', int or array of int
+        Number of values per feature.
+        'auto' : determin feature range from training data.
+        int : maximum value for all features.
+        array : maximum value per feature.
+
+
+    Attributes
+    ----------
+    `feature_indices_` : array of shape (n_features,)
+        Indices to feature ranges. Feature ``i`` in the
+        original data is mapped to features
+        ``feature_indices_[i]`` to ``feature_indices_[i+1]``.
+    `n_values_` : array of shape (n_features,)
+        Maximum number of values per feature.
+
+
+    Examples
+    --------
+    Given a dataset with three features and two
+    data points, we find the maximu value per feature
+    and transform the data to a binary one-hot encoding.
+
+    >>> from sklearn.preprocessing import OneHotEncoder
+    >>> enc = OneHotEncoder()
+    >>> enc.fit([[1, 2, 1]])
+    OneHotEncoder(n_values='auto')
+    >>> enc.n_values_
+    array([2, 3, 2])
+    >>> enc.feature_indices_
+    array([0, 2, 5, 7])
+    >>> enc.transform([[0, 1, 1]]).toarray()
+    array([[1, 0, 0, 1, 0, 0, 1]])
+
+    See also
+    --------
+    LabelEncoder, performs a one-hot encoding on labels.
+    DictVectorizer, performs a one-hot encoding of dictionary items.
+    """
+    def __init__(self, n_values="auto"):
+        self.n_values = n_values
+
+    def fit(self, X, y=None):
+        """Fit OneHotEncoder to X.
+
+        Parameters
+        ----------
+        X : array-like, shape=(n_samples, n_feature)
+            Input array of type int.
+        """
+        X, = check_arrays(X, sparse_format='dense', dtype=np.int)
+        n_features = X.shape[1]
+        if self.n_values == 'auto':
+            n_values = np.max(X, axis=0) + 1
+        elif isinstance(self.n_values, numbers.Integral):
+            n_values = np.empty(n_features, dtype=np.int)
+            n_values.fill(self.n_values)
+        else:
+            try:
+                n_values = np.asarray(self.n_values, dtype=int)
+            except (ValueError, TypeError):
+                raise ValueError("Wrong type for parameter `n_values`."
+                        " Expected 'auto', int or array of ints, got %s"
+                        % type(X))
+            if n_values.shape[0] != X.shape[1]:
+                raise ValueError("Shape mismatch: if n_values is "
+                        "an array, it has to be of shape (n_features,).")
+        self.n_values_ = n_values
+        n_values = np.hstack([[0], n_values])
+        self.feature_indices_ = np.cumsum(n_values)
+        return self
+
+    def transform(self, X):
+        """Transform X using one-hot encoding.
+
+        Parameters
+        ----------
+        X : array-like, shape=(n_samples, feature_indices_[-1])
+            Input array of type int.
+
+        Returns
+        -------
+        X_out : sparse matrix, dtype=int
+            Transformed input.
+        """
+        X, = check_arrays(X, sparse_format='dense', dtype=np.int)
+        n_samples, n_features = X.shape
+
+        n_values_check = np.max(X, axis=0) + 1
+        if (n_values_check > self.n_values_).any():
+            raise ValueError("Feature out of bounds. Try setting n_values.")
+
+        indices = self.feature_indices_
+        if n_features != indices.shape[0] - 1:
+            raise ValueError("X has different shape then during fitting."
+                             " Expected %d, got %d."
+                             % (indices.shape[0] - 1, n_features))
+        features = X + indices[:-1]
+        column_indices = features.ravel()
+        row_indices = np.hstack([i * np.ones(n_features)
+                                    for i in range(n_samples)])
+        data = np.ones(n_samples * n_features)
+        out = sp.coo_matrix((data, (row_indices, column_indices)),
+                shape=(n_samples, indices[-1]), dtype=np.int)
+        return out.tocsr()
 
 
 class LabelEncoder(BaseEstimator, TransformerMixin):